exif/requirements.txt
Pillow
exifread
iptcinfo3
pillow-heif